+++ Seeds of Apocrypha/src/Entities/UI/Button.h
@@ -12,8 +12,8 @@ public:
     }
 
     void Draw() override;
-    void MoveBy(sf::Vector2f offset) override { Entity::MoveBy(offset); Move(); }
-    void MoveTo(sf::Vector2f new_pos) override { Entity::MoveTo(new_pos); Move(); }
+    void MoveBy(sf::Vector2i offset) override { Entity::MoveBy(offset); Move(); }
+    void MoveTo(sf::Vector2i new_pos) override { Entity::MoveTo(new_pos); Move(); }
 
 private:
     //Functions
