+++ Seeds of Apocrypha/src/Entities/UI/Picker.h
@@ -10,8 +10,8 @@ public:
 
     void GetInput() override;
     void Draw() override;
-    void MoveBy(sf::Vector2f offset) override { Entity::MoveBy(offset); Move(); }
-    void MoveTo(sf::Vector2f new_pos) override { Entity::MoveTo(new_pos); Move(); }
+    void MoveBy(sf::Vector2i offset) override { Entity::MoveBy(offset); Move(); }
+    void MoveTo(sf::Vector2i new_pos) override { Entity::MoveTo(new_pos); Move(); }
 
     void SetPicking(const string new_p);
     string GetPicking() { return picking.getString(); }
